pkg/derivation/store: mv {memory,store_test}.go, test all stores
This refactors memory_test.go a bit, so it can test all stores
implementing the derivation.Store interface.
